rename Input
